refactor: 배열 - 에라토스테네스 체
코드 가독성좋게 변경

+++ src/test/java/practice/array/SieveOfEratosthenesTest.java
@@ -28,7 +28,7 @@ public class SieveOfEratosthenesTest {
 
     @MethodSource
     @ParameterizedTest
-    void solution(final long value, final long expected) throws Exception {
+    void solution(final int value, final long expected) throws Exception {
         final long actual = solve.solution(value);
         assertThat(actual).isEqualTo(expected);
     }
